requirements.txt
alembic==1.8.1
click==8.1.3
dataclasses-json==0.5.7
flake8==4.0.1
Inject==4.3.1
mypy==0.971
psycopg2==2.9.3
pytest==7.1.2
pytest-dotenv==0.5.2
pytest-mock==3.8.2
python-dotenv==0.20.0
SQLAlchemy==1.4.39
fastapi==0.79.0
uvicorn==0.18.2
pydantic==1.9.2
starlette==0.19.1